requirements.txt
python-pptx
requests
Pillow
